New activities Jutge.org (Bucles)

+++ rounding.cc
@@ -2,14 +2,14 @@
 
 int main()
 {
-  float a=0;
+  double a=0;
   std::cin >> a;
 
-  if( (a-(int)a)==0 )
-    std::cout << (int)a << " " << (int)a << " " << (int)a << '\n';
-  else if( (a-(int)a)>=0.5 )
-    std::cout << (int)a << " " << ((int)a+1) << " " << ((int)a+1) <<'\n';
+  if( (a-(unsigned long)a)==0 )
+    std::cout << (unsigned long)a << " " << (unsigned long)a << " " << (unsigned long)a << '\n';
+  else if( (a-(double)((unsigned long)a))>=0.5 )
+    std::cout << (unsigned long)a << " " << ((unsigned long)a+1) << " " << ((unsigned long)a+1) <<'\n';
   else
-    std::cout << (int)a << " " << ((int)a+1) << " " << (int)a <<'\n';
+    std::cout << (unsigned long)a << " " << ((unsigned long)a+1) << " " << (unsigned long)a <<'\n';
   return 0;
 }
